MMRP/requirements.txt
PyYAML>=6.0
Pillow>=10.0.0
psutil>=5.8.0
opencv-python>=4.7.0.72
numpy>=1.24.2
blobfile>=2.0.1
tqdm>=4.65.0